+++ CHANGELOG.md
@@ -1,9 +1,19 @@
 # Change Log
 
-All notable changes to the "typst--" extension will be documented in this file.
+All notable changes to the "typstf" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.0.6] - 2026-04-02
 
-## [Unreleased]
+### Added
+- Added Markdown export feature: Convert Typst documents to Markdown format
+- Support for converting Typst formulas to LaTeX format
+- Support for headings, bold, italic, links, images, and lists conversion
 
+## [0.0.5] - 2025-XX-XX
+
+### Added
 - Initial release
+- Export inline formulas to SVG
+- Support for selected text export
+- Auto-naming with SHA256 hash
+- Right-click menu integration